fix: replacing create-session modal
