+++ ui/src/features/goals/components/goals.tsx
@@ -22,7 +22,6 @@ import {
 import { useQuery } from '@tanstack/react-query';
 import { useRouteContext, useNavigate } from '@tanstack/react-router';
 
-import CreateSessionModal from '@/features/goal/components/create-session-modal';
 import { getStudentGoalsOptions } from '../api';
 import { statusMapper } from '../mappers';
 import type { Goal } from '../models';
@@ -67,9 +66,6 @@ export default function Goals() {
       </Header>
 
       <Container py="xl" mx="xl" fluid>
-        <Group justify="flex-end">
-          <CreateSessionModal />
-        </Group>
         <GoalsList goals={goals} isLoading={isLoading} error={error} />
       </Container>
     </>
